Move options and add new fast test for making sure things actually work
