+++ src/toil/batchSystems/tes.py
@@ -35,8 +35,8 @@ from requests.exceptions import HTTPError
 from toil import applianceSelf
 from toil.batchSystems.abstractBatchSystem import (EXIT_STATUS_UNAVAILABLE_VALUE,
                                                    BatchJobExitReason,
-                                                   OptionSetter,
                                                    UpdatedBatchJobInfo)
+from toil.batchSystems.options import OptionSetter
 from toil.batchSystems.cleanup_support import BatchSystemCleanupSupport
 from toil.batchSystems.contained_executor import pack_job
 from toil.common import Config, Toil
